WIKI-1387 give priority to page ACL instead of wiki ACL (#149)
The wiki permission is used only to check if user has access to page. This is problematic since the user will not be able to access pages if he is not allowed to view wiki.
This PR combine the checks about wiki access and page access to see if user can access the page.