WIKI-1294 : save edition in progress even for the first draft of a user for a given page
When an user edits a page and updates some content, when the draft is saved (after 30 seconds by default), we save that the page is being edited by the user.
This is done correctly, except for the first draft (the first draft after an user edited a page content).
This fix do it in all cases (not only the first draft).