WIKI-1392 : encode HTML contents at display time only
HTML contents (like activities contents) must be interpreted as HTML (so they should be sanitized but not encoded).
Non-HTML contents (like wiki page title, wiki template page title, ...) must not be interpreted as HTML (so they should be encoded).
This fix :
* do not encode page title when storing it in the database (stored as is in the database)
* encode wiki pages title in wiki tree when displaying it (instead of sanitizing them)

So if a page has the title "<b>Page 1</b>", it will be displayed as is on tp the page and in the wiki pages tree.