MAINT-25355 : refresh UIExtensions before executing process Action (#259)
When opening a page to Edit and then opening a parent page to view, the update wiki page operation fails.
The the UI extensions used to save the page will be updated by the view of parent page and won't be rendered, thus when executing the Save action, we will get a NPE as the Save action of the subcomponent will be missing.
The fix refreshes the UIExtensions attached to each UIComponent when an action is fired to make sure we have all sub components available/updated with their actions.

(cherry picked from commit ba26e8e93d010a1fa5957e79bd0f069819a56622)